requirements.txt
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
tiktoken>=0.5.0
tabulate
sentence-transformers>=2.2.0
scikit-learn
aiohttp>=3.8.0
numpy>=1.24.0
jupyter>=1.0.0
pytest>=7.0.0
tabulate
# Remove existing virtual environments
#rm -rf venv/
#rm -rf research/